% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{extract_params}
\alias{extract_params}
\title{Extract parameters from table_builder or use direct arguments}
\usage{
extract_params(data_or_builder, strata, output, round, data_override = NULL)
}
\description{
Extract parameters from table_builder or use direct arguments
}
\keyword{internal}
