% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_availability}
\alias{get_availability}
\title{Get N(\%) availability - the number of times the variable is not NA}
\usage{
get_availability(data, strata, variable, name, output, round = 2)
}
\arguments{
\item{data}{Either a tibble OR a table_builder object from make_output_df}

\item{strata}{variable to stratify output by. Needs to be a factor to force to ordering to work. (Optional if using table_builder object)}

\item{variable}{Variable to check availability for. Can be numeric, string, factor.}

\item{name}{string name to put into the row}

\item{output}{The dataframe to append the requested summary to (Optional if using table_builder object)}

\item{round}{The number of decimal places to round results to. (Optional if using table_builder object)}
}
\description{
Get N(\%) availability - the number of times the variable is not NA
}
