% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_count}
\alias{get_count}
\title{Count number of non-missing values
Displays number of non-missing rows.}
\usage{
get_count(data, strata, variable, name, output)
}
\arguments{
\item{data}{tibble containing data. Can't be a grouped tibble}

\item{output}{}
}
\value{

}
\description{
Count number of non-missing values
Displays number of non-missing rows.
}
