% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_mean_sd}
\alias{get_mean_sd}
\title{Title Get mean and SD for a particular variable. By group.
Does a one-way ANOVA test (t-test if only 2 groups) if there the output dataframe had a p value column.}
\usage{
get_mean_sd(data, strata, variable, name, output)
}
\arguments{
\item{data}{data frame containing required data}

\item{strata}{variable to stratify output by. Needs to be a factor to force to ordering to work.}

\item{variable}{name of variable to be summarised}

\item{name}{string name to put in the row.}

\item{output}{}
}
\description{
Title Get mean and SD for a particular variable. By group.
Does a one-way ANOVA test (t-test if only 2 groups) if there the output dataframe had a p value column.
}
