% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{compute_smd}
\alias{compute_smd}
\title{Compute SMD for a variable against strata
Returns formatted SMD string, or "" if it fails.}
\usage{
compute_smd(data, variable, strata)
}
\description{
Compute SMD for a variable against strata
Returns formatted SMD string, or "" if it fails.
}
\keyword{internal}
