% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_availability}
\alias{get_availability}
\title{Get N(\%) availability - the number of times the variable is not NA}
\usage{
get_availability(data, strata, variable, name, output, round = 2)
}
\arguments{
\item{data}{tibble containing data. Can't be a grouped tibble}

\item{strata}{variable to stratify output by. Needs to be a factor to force to ordering to work.}

\item{variable}{Variable to check availability for. Can be numeric, string, factor.}

\item{name}{string name to put into the row}

\item{output}{The dataframe to append the requested summary to}

\item{round}{The number of decimal places to round results to.}
}
\description{
Get N(\%) availability - the number of times the variable is not NA
}
