% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_count}
\alias{get_count}
\title{Count number of non-missing values
Displays number of non-missing rows.}
\usage{
get_count(data, strata, variable, name, output)
}
\arguments{
\item{data}{tibble containing data. Can't be a grouped tibble}

\item{strata}{variable to stratify output by. Needs to be a factor to force to ordering to work.}

\item{variable}{Ideally a factor variable. If not, gets converted to factor anyway.}

\item{name}{string name to put into the row}

\item{output}{The dataframe to append the requested summary to}
}
\description{
Count number of non-missing values
Displays number of non-missing rows.
}
