% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_median_iqr}
\alias{get_median_iqr}
\title{Title Get Median and IQR for a particular variable. By group.}
\usage{
get_median_iqr(data, strata, variable, name, output)
}
\arguments{
\item{data}{data frame containing required data}

\item{strata}{variable to stratify output by. Needs to be a factor to force to ordering to work.}

\item{variable}{name of variable to be summarised}

\item{name}{string name to put in the row.}

\item{output}{}
}
\description{
Title Get Median and IQR for a particular variable. By group.
}
