% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_n_percent}
\alias{get_n_percent}
\title{Get count and percentage. By group.}
\usage{
get_n_percent(
  data,
  strata,
  variable,
  name,
  output,
  id = "",
  round = 2,
  sort_by_freq = FALSE,
  data_override = NULL
)
}
\arguments{
\item{data}{Either a tibble OR a table_builder object from make_output_df}

\item{strata}{variable to stratify output by. Needs to be a factor to force to ordering to work. (optional if using table_builder object)}

\item{variable}{Ideally a factor variable. If not, gets converted to factor anyway.}

\item{name}{string name to put into the row}

\item{output}{The dataframe to append the requested summary to (optional if using table_builder object)}

\item{id}{Optional. Character vector containing name of variable to use when counting the denominator. Allows the sum of the numerators to be greater than 100 if there is more than one row per denominator variable.}

\item{round}{The number of decimal places to round results to (optional if using table_builder object)}

\item{sort_by_freq}{TRUE/FALSE for whether to overwrite factor ordering and sort in descending order of total frequency.}

\item{data_override}{Override data in table_builder object with this data frame (optional, rarely used)}
}
\description{
Gets count and percentage for factor variables by group.
NAs and empty strings are converted to a separate level "Missing"
Does a chisquare test if the 'output' argument has a p value column in it.
Unless ID variable is specified, the denominator is the number of rows in the dataset.
}
