% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_n_percent}
\alias{get_n_percent}
\title{Gets count and percentage for factor variables by group. Does a chisquare test if the 'output' argument has a p value column in it.}
\usage{
get_n_percent(data, strata, variable, name, output)
}
\arguments{
\item{data}{tibble containing data. Can't be a grouped tibble}

\item{output}{}
}
\value{

}
\description{
Gets count and percentage for factor variables by group. Does a chisquare test if the 'output' argument has a p value column in it.
}
