% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_sum}
\alias{get_sum}
\title{Get sum of a numeric variable.
Treats missing values as zero.}
\usage{
get_sum(data, strata, variable, name, output, round = 2)
}
\arguments{
\item{data}{Either a tibble OR a table_builder object from make_output_df}

\item{strata}{variable to stratify output by. Needs to be a factor to force to ordering to work. (Optional if using table_builder object)}

\item{variable}{Numeric variable to get sum for}

\item{name}{string name to put into the row}

\item{output}{The dataframe to append the requested summary to (Optional if using table_builder object)}
}
\description{
Get sum of a numeric variable.
Treats missing values as zero.
}
