% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_unique_count}
\alias{get_unique_count}
\title{Count number of unique non-missing values
Displays number of non-missing rows.}
\usage{
get_unique_count(data, strata, variable, name, output)
}
\arguments{
\item{data}{Either a tibble OR a table_builder object from make_output_df}

\item{strata}{variable to stratify output by. Needs to be a factor to force to ordering to work. (Optional if using table_builder object)}

\item{variable}{Ideally a factor variable. If not, gets converted to factor anyway.}

\item{name}{string name to put into the row}

\item{output}{The dataframe to append the requested summary to (Optional if using table_builder object)}
}
\description{
Count number of unique non-missing values
Displays number of non-missing rows.
}
