% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{make_output_df}
\alias{make_output_df}
\title{Make output dataframe}
\usage{
make_output_df(data, strata, include_tests = FALSE)
}
\arguments{
\item{data}{tibble containing data. Can't be a grouped tibble. Also, can't have a variable named 'variable'.}

\item{strata}{A factor variable to stratify by.}

\item{include_tests}{Should it create a space for p-values?}
}
\description{
Includes neat column names with counts of observations in each group.
}
